requirements.txt
google-genai
fastapi
uvicorn
openai
requests
youtube-transcript-api